Fractal Patterns By Using Koch Curves

+++ 9_Fractal.cpp
@@ -12,8 +12,8 @@ void koch(int it,int x1,int y1,int x2,int y2){
         dy = (y2-y1)/3;
         x3 = x1+dx;
         y3 = y1+dy;
-        x4 = (int)(0.5*(x1+x2)+sqrt(3)*(y1-y2)/6);
-        y4 = (int) (0.5*(y1+y2)+sqrt(3)*(x2-x1)/6);
+        x4 = 0.5*(x1+x2)+sqrt(3)*(y1-y2)/6;
+        y4 = 0.5*(y1+y2)+sqrt(3)*(x2-x1)/6;
         x5 = 2*dx+x1;
         y5 = 2*dy+y1;
         koch(it-1,x1,y1,x3,y3);
